python_backend/requirements.txt
flask==3.0.0
pymongo==4.6.0
flask-login==0.6.3
python-dotenv==1.0.0
bcrypt==4.0.1
python-jose==3.3.0
gunicorn==21.2.0
flask-cors==4.0.0